requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.3
ollama==0.1.5
pyyaml==6.0.1
python-multipart==0.0.6
Levenshtein==0.21.1
nltk==3.8.1
requests==2.31.0